+++ src/enums/BorderMethod.php
@@ -6,5 +6,8 @@ enum BorderMethod: string
 {
 	case OVERLAY = 'overlay';
 	case SHRINK = 'shrink';
+	case EXPAND = 'expand';
+
+	/** @deprecated Use EXPAND instead. */
 	case PAD = 'pad';
 }